requirements.txt
APScheduler==3.9.1
Django==4.0.6
PyJWT==2.4.0
pytz==2022.1
SQLAlchemy==1.4.37
websockets==10.3